Move standard_climo_periods() to __init__.py
