+++ nchelpers/util.py
@@ -22,19 +22,6 @@ def resolution_standard_name(seconds):
     }.get(seconds, 'other')
 
 
-def standard_climo_periods(calendar='standard'):
-    standard_climo_years = {
-        '6190': ['1961', '1990'],
-        '7100': ['1971', '2000'],
-        '8110': ['1981', '2010'],
-        '2020': ['2010', '2039'],
-        '2050': ['2040', '2069'],
-        '2080': ['2070', '2099']
-    }
-    day = '30' if calendar == '360_day' else '31'
-    return dict([(k, (s2d(year[0]+'-01-01'), s2d(year[1]+'-12-'+day))) for k, year in standard_climo_years.items()])
-
-
 def time_to_seconds(x, units='seconds'):
     '''Return the number of seconds equal to `x` `units` of time, e.g., 10 minutes'''
     seconds_per_unit = {
